// src/model/file/TexFileInfoModel.ts
export interface TexFileInfoModel { 
    id: string; 
    name: string;
    created_time: string;
    updated_time: string;
    user_id: string;
    nickname: string;
    doc_status: number;
    project_id: string;
    file_type: number;
    file_id: string;
    parent: string;
    main_flag: number;
    sort: number;
    yjs_initial: number;
    file_path: string;
}